LRNG: mark LRNG non-operation if init DRNG is not seeded
When a DRNG becomes not fully seeded, the LRNG will still happily
continue unless the DRNG is the initial DRNG. The reason is that when a
DRNG reference is to be obtained, the LRNG always uses the
NUMA-node-local instance if it is fully seeded. If it is not, the
initial DRNG instance is used instead.

Thus, only when the initial DRNG instance becomes not fully seeded, the
LRNG is marked as non-operational as there is no other fallback.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: fdlyfz <fadlidodo2004@gmail.com>