LRNG: calculate GCD of timer used for IRQ ES
The IRQ ES uses the 8 LSB of the event time stamp added to the entropy
pool. It is possible, however, that the time stamp moves in increments
other than 1. For example, AMD Ryzen 5950X has always the least
significant bit unset. Other CPUs are observed to even have more LSBs
being unchanged, such as certain IBM System Z processors. Finally, it is
possible that time stamps are made coarse by hypervisors by masking out
some least significant bits of the time stamp.

As the LRNG uses the 8 LSB of the time stamp, any non-moving bits should
be ignored as only moving bits can provide entropy. The GCD calculation
applied to the first time stamps can detect such non-moving bits which
are equal to the GCD of the analyzed time stamps. Once the analysis is
complete, the LRNG uses the GCD value on each time stamp obtained for an
IRQ.

On the AMD Ryzen 5950X test system, the GCD is larger than 1:

[    0.948771] lrng_es_irq: Setting GCD to 2

For this CPU therefore, the least significant bit is now eliminated by
means of the division.

When the GCD calculation is not yet completed, the entire 32 bit of the
time stamp is inserted into the pool. When the GCD is obtained, only the
8 LSB of the time stamp after applying the GCD are inserted into the
pool.

If the LRNG is reset, the GCD is also reset which implies that the next
time stamps are used for the GCD calculation.

AS the correctness of the GCD is supportive to a high quality entropy
(read a higher entropy rate), the self-test code is extended to execute
a GCD self test.

Note, the GCD calculation is only applied to the first time stamps after
reset including boot. At runtime, the GCD logic does not affect the
performance.

A test verifying the operation of the GCD is provided.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: fdlyfz <fadlidodo2004@gmail.com>